refactor(messaging): Move `*Signed` types into an `authority` module
This sets up the contained structures for more general use that within
`MsgKind`.
