+++ src/messaging/msg_kind.rs
@@ -6,12 +6,8 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use super::node::{KeyedSig, SigShare};
-use crate::types::{PublicKey, Signature};
-use bls::PublicKey as BlsPublicKey;
-use ed25519_dalek::{PublicKey as EdPublicKey, Signature as EdSignature};
+use super::{BlsShareSigned, ClientSigned, NodeSigned, SectionSigned};
 use serde::{Deserialize, Serialize};
-use xor_name::XorName;
 
 /// Source authority of a message.
 ///
@@ -50,48 +46,4 @@ pub enum MsgKind {
     /// A message from an Elder node with authority of its whole section.
     // FIXME: find an example.
     SectionSignedMsg(SectionSigned),
-}
-
-/// Authority of a client
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-pub struct ClientSigned {
-    /// Client public key.
-    pub public_key: PublicKey,
-    /// Client signature.
-    pub signature: Signature,
-}
-
-/// Authority of a single peer.
-#[derive(Clone, Eq, PartialEq, Serialize, Deserialize, custom_debug::Debug)]
-pub struct NodeSigned {
-    /// Section key of the source.
-    pub section_pk: BlsPublicKey,
-    /// Public key of the source peer.
-    #[debug(with = "PublicKey::fmt_ed25519")]
-    pub public_key: EdPublicKey,
-    /// Ed25519 signature of the message corresponding to the public key of the source peer.
-    #[debug(with = "Signature::fmt_ed25519")]
-    pub signature: EdSignature,
-}
-
-/// Authority of a single peer that uses it's BLS Keyshare to sign the message.
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-pub struct BlsShareSigned {
-    /// Section key of the source.
-    pub section_pk: BlsPublicKey,
-    /// Name in the source section.
-    pub src_name: XorName,
-    /// Proof Share signed by the peer's BLS KeyShare.
-    pub sig_share: SigShare,
-}
-
-/// Authority of a whole section.
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-pub struct SectionSigned {
-    /// Section key of the source.
-    pub section_pk: BlsPublicKey,
-    /// Name in the source section.
-    pub src_name: XorName,
-    /// BLS proof of the message corresponding to the source section.
-    pub sig: KeyedSig,
 }
